CASE Model

Perform Multi-trait Fine-mapping

CASE(Z = NULL, R, hatB = NULL, hatS = NULL, N, V = NULL, cs = TRUE, ...)

Arguments

Z

M * C matrix of z scores.

R

M * M matrix of LD.

hatB

M * C matrix of the estimated effects. Alternative summary data (together with hatS) to be provided instead of Z.

hatS

M * C matrix of standard errors of the estimated effects. Alternative summary data (together with hatB) to be provided instead of Z.

N

either C vector of the sample size, or C * C matrix of the sample size (diagonal) and ovelaps (off-diagonal). If provided with a vector, CASE assumes that each pair of traits overlaps with their minimal sample size.

V

(optional) C * C covariance (correlation) matrix for the noise between traits. If not provided, the default is an identity matrix.

cs

logical, whether to get credible sets.

...

additional arguments.

Value

A "CASE" object with the following elements:

pi:

L-vector, the prior probabilities of sharing patterns.

U:

L-list of C * C matrix, the prior covariances of sharing patterns.

V:

C * C matrix, the sample-adjusted phenotypical variance.

Details

TBD

References

TBD

Author

Chen Lin, Hongyu Zhao

Examples

set.seed(3)
N = 500
M = 1000
C = 3
X = matrix(sample(0:2, size = N * M, replace = TRUE), N, M)
B = matrix(0, M, C)
B[1, ] = runif(C, 1, 2)
B[2, 1] = runif(1, 2, 3)
e = matrix(rnorm(N * C), N, C)
Y = X %*% B + e

X = scale(X)
Y = scale(Y)
R = cor(X)

Z = hatB = hatS = matrix(0, M, 3)
for (c in 1:3){
  hatB[, c] <- t(X) %*% Y[, c] / (N - 1)
  hatS[, c] <- sqrt((sum(Y[, c]^2)  - hatB[, c]^2 * (N - 1)) / N / (N - 2))
  Z[, c] <- hatB[, c] / hatS[, c]
}

fit <- CASE(Z = Z, R = R, N = rep(N, 3))
#> Initialize with patterns:  111 100 010 001 000 
#> Start Posterior Analysis.Start getting credible sets.
print(fit$sets)
#> [[1]]
#> [[1]]$cs
#> [[1]]$cs[[1]]
#> [1] 1
#> 
#> [[1]]$cs[[2]]
#> [1] 2
#> 
#> 
#> [[1]]$purity_min_cor
#> [1] 1 1
#> 
#> [[1]]$coverage
#> [1] 1 1
#> 
#> 
#> [[2]]
#> [[2]]$cs
#> [[2]]$cs[[1]]
#> [1] 1
#> 
#> 
#> [[2]]$purity_min_cor
#> [1] 1
#> 
#> [[2]]$coverage
#> [1] 1
#> 
#> 
#> [[3]]
#> [[3]]$cs
#> [[3]]$cs[[1]]
#> [1] 1
#> 
#> 
#> [[3]]$purity_min_cor
#> [1] 1
#> 
#> [[3]]$coverage
#> [1] 1
#> 
#>